# frontend/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crypto Radar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/exchanges/connections": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "List Exchange Connections",
        "operationId": "list_exchange_connections_api_v1_exchanges_connections_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo_user",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExchangeConnectionResponse"
                  },
                  "title": "Response List Exchange Connections Api V1 Exchanges Connections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exchanges"
        ],
        "summary": "Create Exchange Connection",
        "operationId": "create_exchange_connection_api_v1_exchanges_connections_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeConnectionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exchanges/connections/{connection_id}": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Get Exchange Connection",
        "operationId": "get_exchange_connection_api_v1_exchanges_connections__connection_id__get",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "exchanges"
        ],
        "summary": "Update Exchange Connection",
        "operationId": "update_exchange_connection_api_v1_exchanges_connections__connection_id__patch",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeConnectionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "exchanges"
        ],
        "summary": "Delete Exchange Connection",
        "operationId": "delete_exchange_connection_api_v1_exchanges_connections__connection_id__delete",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exchanges/connections/{connection_id}/test": {
      "post": {
        "tags": [
          "exchanges"
        ],
        "summary": "Test Exchange Connection",
        "operationId": "test_exchange_connection_api_v1_exchanges_connections__connection_id__test_post",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeConnectionActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exchanges/connections/{connection_id}/sync": {
      "post": {
        "tags": [
          "exchanges"
        ],
        "summary": "Sync Exchange Connection Trades",
        "operationId": "sync_exchange_connection_trades_api_v1_exchanges_connections__connection_id__sync_post",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealTradeImportResult"
                }
              }
            }
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealTradeImportNotReadyResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExchangeConnectionActionResponse": {
        "properties": {
          "connection": {
            "$ref": "#/components/schemas/ExchangeConnectionResponse"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "stubbed"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "connection",
          "message"
        ],
        "title": "ExchangeConnectionActionResponse"
      },
      "ExchangeConnectionCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "demo_user"
          },
          "exchange_code": {
            "type": "string",
            "minLength": 1,
            "title": "Exchange Code"
          },
          "label": {
            "type": "string",
            "minLength": 1,
            "title": "Label"
          },
          "account_type": {
            "type": "string",
            "title": "Account Type",
            "default": "spot"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Secret"
          },
          "api_passphrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Passphrase"
          },
          "permissions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Permissions"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "exchange_code",
          "label"
        ],
        "title": "ExchangeConnectionCreateRequest"
      },
      "ExchangeConnectionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "exchange_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exchange Id"
          },
          "exchange_code": {
            "type": "string",
            "title": "Exchange Code"
          },
          "exchange_name": {
            "type": "string",
            "title": "Exchange Name"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "account_type": {
            "type": "string",
            "title": "Account Type"
          },
          "key_ref": {
            "type": "string",
            "title": "Key Ref"
          },
          "permissions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Permissions"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "exchange_id",
          "exchange_code",
          "exchange_name",
          "label",
          "account_type",
          "key_ref",
          "permissions",
          "status",
          "last_sync_at",
          "metadata",
          "created_at"
        ],
        "title": "ExchangeConnectionResponse"
      },
      "ExchangeConnectionUpdateRequest": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "account_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Type"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Secret"
          },
          "api_passphrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Passphrase"
          },
          "permissions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permissions"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "ExchangeConnectionUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RealTradeImportNotReadyResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "not_implemented",
            "title": "Status",
            "default": "not_implemented"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "connection_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          },
          "connector_required": {
            "type": "boolean",
            "title": "Connector Required",
            "default": true
          },
          "normalized_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Normalized Targets"
          },
          "analytics_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Analytics Targets"
          },
          "raw_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Raw Targets"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "RealTradeImportNotReadyResponse"
      },
...
      "RealTradeImportResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "dry_run"
            ],
            "title": "Status"
          },
          "connection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Connection Id"
          },
          "external_exchange_orders_written": {
            "type": "integer",
            "title": "External Exchange Orders Written",
            "default": 0
          },
          "external_exchange_trades_written": {
            "type": "integer",
            "title": "External Exchange Trades Written",
            "default": 0
          },
          "analytics_events_written": {
            "type": "integer",
            "title": "Analytics Events Written",
            "default": 0
          },
          "raw_events_written": {
            "type": "integer",
            "title": "Raw Events Written",
            "default": 0
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "connection_id",
          "started_at",
          "finished_at"
        ],
        "title": "RealTradeImportResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}